requirements.txt
torch==1.9.0+cu111
torchvision==0.10.0+cu111
cupy==8.6.0
numpy==1.23.3
torchinfo==1.7.0
einops==0.6.0
matplotlib==3.6.0
pillow==9.2.0
scikit-learn==1.1.2
tqdm==4.64.1  # for progress bars during training/evaluation